Move moments/test_moments_rolling_apply.py to window/test_apply.py
